% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{char_classify}
\alias{char_classify}
\title{char_classify}
\usage{
char_classify(data, rule)
}
\arguments{
\item{data}{a character vector of data to be tested against a rule.}

\item{rule}{a character vector of length one that specifies the rule to test.}
}
\value{
logical vector indicating which elements of the data match the rule.
}
\description{
char_classify
}

