% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{character_classify}
\alias{character_classify}
\title{character_classify}
\usage{
character_classify(var_name, data, rule)
}
\arguments{
\item{var_name}{the current varaible under test #!.}

\item{data}{data.frame whose rows are to be categorised according to a provided set of rules.}

\item{rule}{a character vector of length one that specifies the rule to test.}
}
\value{
logical vector indicating which elements of the data match the rule.
}
\description{
character_classify
}
