% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{check_definitions}
\alias{check_definitions}
\title{check_definitions}
\usage{
check_definitions(data, rules)
}
\arguments{
\item{definitions}{a data.frame}
}
\value{
nada
}
\description{
check_definitions
}

