% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_dictionary}
\alias{check_dictionary}
\title{check_dictionary}
\usage{
check_dictionary(dictionary)
}
\arguments{
\item{dictionary}{Each row of the data frame dictionary specifies a data type for a variable
found in data and rules. There should be two columns in dictionary, variable and type.
The column names in rules should be exactly matched in dictionary and type should be one of num,
char, or date for each variable.}
}
\description{
check_dictionary
}

