% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_rules}
\alias{check_rules}
\title{check_rules}
\usage{
check_rules(data, rules)
}
\arguments{
\item{data}{data.frame whose rows are to be classified according to a set of rules.}

\item{rules}{The first column of this data frame should be the name of the (to be created)
classification variable and subsequent columns should be the names of the variables which
are to be tested against a rule.
Each row of the rules data.frame should contain a classification level in the first column and
from there the rules to test each column against.}
}
\description{
check_rules
}

