% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_type}
\alias{check_type}
\title{check_type}
\usage{
check_type(type_to_check, data, dictionary)
}
\arguments{
\item{type_to_check}{the data type to check (numeric, character, date)}

\item{data}{data.frame whose rows are to be categorised according to a provided set of rules.}

\item{dictionary}{A data frame used by the function classify.
Each row of the data frame dictionary specifies a data type for a variable
found in data and rules. There are two columns in dictionary, variable and type.
The column names in rules are matched in dictionary and type is one of numeric,
character, or date for each variable.}
}
\description{
check_type
}

