% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{classify}
\usage{
classify(data, dictionary, rules, default_def = "unknown")
}
\arguments{
\item{data}{data.frame whose rows are to be classified according to a set of rules.}

\item{dictionary}{data.frame which specifies the data type of the variables in data which are
tested against a rule in rules.}

\item{rules}{data.frame. The first column should be the name of the classification variable and
subsequent columns should be the names of the variables which are to be tested against a rule.
Each row of the rules data.frame should contain a classification level in the first column and
from there the rules to test each column against.}

\item{default_def}{the default value to be given to all unclassified rows.}
}
\value{
data.frame with added first column which classifies the data according to a set of rules.
}
\description{
classify
}

