% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{numeric_classify}
\alias{numeric_classify}
\title{numeric_classify}
\usage{
numeric_classify(var_name, data, rule)
}
\arguments{
\item{var_name}{the current varaible under test #!.}

\item{data}{data.frame whose rows are to be categorised according to a provided set of rules.}

\item{rule}{a character vector of length one that specifies the rule to test. The rule can either
indicate a value must be present ("1") or absent ("0") or test the entered value using one of the
set of comparison functions ("==", ">=", "<=", ">", "<"). If a comparison function is used any NA
value in the data is assumed to have failed the test. If an empty string is entered for the rule,
any value is assumed correct.}
}
\value{
logical vector indicating which elements of the data match the rule.
}
\description{
numeric_classify
}
