% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_dictionary}
\alias{check_dictionary}
\title{check_dictionary}
\usage{
check_dictionary(dictionary)
}
\arguments{
\item{dictionary}{a data.frame with columns variable and type}
}
\description{
check_dictionary
}

