% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_rules}
\alias{check_rules}
\title{check_rules}
\usage{
check_rules(data, rules)
}
\arguments{
\item{data}{data.frame whose rows are to be categorised according to a provided set of rules.}

\item{rules}{rules #!.}
}
\description{
check_rules
}
