% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{date_classify}
\alias{date_classify}
\title{date_classify}
\usage{
date_classify(data, rule)
}
\arguments{
\item{data}{a date vector of data to be tested against a rule.}

\item{rule}{a character vector of length one that specifies the rule to test. For date data only
present (1) or absent (0) rules are testable. If an empty string is entered for the rule, any
value is assumed correct.}
}
\value{
logical vector indicating which elements of the data match the rule.
}
\description{
date_classify
}

